isolating station specific deps from client specific deps using separate modules
