+++ go.work
@@ -1,0 +1,4 @@
+use (
+	.
+	./pkg/regserver
+)
