+++ pkg/regserver/apiregserver/apiregserver.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/refraction-networking/conjure/pkg/metrics"
-	"github.com/refraction-networking/conjure/pkg/regprocessor"
+	"github.com/refraction-networking/conjure/pkg/regserver/regprocessor"
 	"github.com/refraction-networking/conjure/pkg/station/lib"
 	pb "github.com/refraction-networking/conjure/proto"
 	log "github.com/sirupsen/logrus"
