+++ pkg/regserver/dnsregserver/dnsregserver_test.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/refraction-networking/conjure/pkg/metrics"
-	"github.com/refraction-networking/conjure/pkg/regprocessor"
+	"github.com/refraction-networking/conjure/pkg/regserver/regprocessor"
 	pb "github.com/refraction-networking/conjure/proto"
 	log "github.com/sirupsen/logrus"
 	"google.golang.org/protobuf/proto"
