+++ pkg/regserver/go.mod
@@ -1,0 +1,3 @@
+module github.com/refraction-networking/conjure/pkg/regserver
+
+go 1.20
