+++ pkg/regserver/regprocessor/auth_test.go
